Sigh, I was going to do it piecemeal, but it became too hard. Check in a number of fixes for header files in particular with prototypes being wrong, many required functions missing, missing types, missing fields etc.
Also changes the method used to protect multiple definitions of a number of
types (to a more useful way IMNSHO).

Fix up the source code broken by the header file changes.


git-svn-id: https://svn.aros.org/svn/aros/trunk@14424 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/clib/gcvt.c
@@ -17,7 +17,7 @@
 
 /*  SYNOPSIS */
 	double	  number,
-	size_t	  ndigit,
+	int	  ndigit,
 	char	* buf
 	)
 
@@ -49,7 +49,7 @@
 
 ******************************************************************************/
 {
-    sprintf (buf, "%.*G", (int)ndigit, number);
+    sprintf (buf, "%.*G", ndigit, number);
 
     return buf;
 } /* sprintf */
